LayoutTests/inspector/console-uncaught-exception-expected.txt

CONSOLE MESSAGE: line 35: Error: Exception in inline script.
CONSOLE MESSAGE: line 19: Error: Exception in 'load' event listener.
CONSOLE MESSAGE: line 27: Error: Exception in setTimeout callback.
CONSOLE MESSAGE: line 35: Error: Exception in inline script.
CONSOLE MESSAGE: line 19: Error: Exception in 'load' event listener.
CONSOLE MESSAGE: line 27: Error: Exception in setTimeout callback.
Tests that uncaught exceptions are logged into console.Bug 47250.

Page reloaded.
console-uncaught-exception.html:35Error: Exception in inline script.
console-uncaught-exception.html:19Error: Exception in 'load' event listener.
console-uncaught-exception.html:27Error: Exception in setTimeout callback.
